nixpkgs: Bootstrap standalone emacs
Awesome doc update from
https://nixos.org/nixos/manual/#idm140737315714640

I now have mostly what i tried to have earlier when asking for emacs +
nix as self-contained (https://github.com/NixOS/nixpkgs/issues/10072).

That is separation of concern between the libraries dependencies to
install and the require statements from the .emacs.d/init.el
routine (which is(/was?) flaky in emacs).

There are still some work to do to remove:
- haskell binaries
- ocaml binaries
- etc...

But this should be a matter of updating the packages (last list to
concatenate at the end of emacs.nix).